Pull non-shared upgrade version of plus promo modal out of shared PlusPromoModal component and into Panel to simplify
