+++ app/hub/Views/HomeView/HomeViewContainer.jsx
@@ -95,7 +95,6 @@ class HomeViewContainer extends Component {
 					show={!isPlus && !plus_promo_modal_shown}
 					location="hub"
 					clickHandler={this._handlePlusPromoModalClicks}
-					version={PlusPromoModal.INITIAL}
 				/>
 				<HomeView {...childProps} />
 			</div>
